完善了异常处理

+++ src/main/java/com/example/server/service/Impl/PoiServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.server.service.Impl;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.example.server.exception.PoiException;
 import com.example.server.mapper.PicMapper;
 import com.example.server.mapper.PoiMapper;
 import com.example.server.pojo.Pic;
@@ -19,32 +20,49 @@ public class PoiServiceImpl extends ServiceImpl<PoiMapper, Poi> implements IPoiS
     private PicMapper picMapper;
     @Override
     public void saveMain(Poi poi, List<Pic> pics) {
-        poiMapper.insert(poi);
+        int rowNum = poiMapper.insert(poi);//返回插入数据数目
 
+        if (rowNum == 0){
+            throw PoiException.operateFail();
+        }
         if (pics != null) {
             for (Pic pic:pics){
                 pic.setPoiId(poi.getId());
-                picMapper.insert(pic);
+                rowNum = picMapper.insert(pic);
+                if (rowNum == 0){
+                    throw PoiException.operateFail();
+                }
             }
         }
     }
 
     @Override
     public void deleteMain(Integer id) {
-        poiMapper.deleteById(id);
-
-        picMapper.deleteByPoiId(id);
+        int rowNum = poiMapper.deleteById(id);
+        if (rowNum == 0){
+            throw PoiException.operateFail();
+        }
+        rowNum = picMapper.deleteByPoiId(id);
+        if (rowNum == 0){
+            throw PoiException.operateFail();
+        }
     }
 
     @Override
     public void updateMain(Poi poi, List<Pic> pics) {
-        poiMapper.updateById(poi);
+        int rowNum = poiMapper.updateById(poi);
 
+        if (rowNum == 0){
+            throw PoiException.operateFail();
+        }
         picMapper.deleteByPoiId(poi.getId());
         if (pics != null) {
             for (Pic pic:pics){
                 pic.setPoiId(poi.getId());
-                picMapper.insert(pic);
+                rowNum = picMapper.insert(pic);
+                if (rowNum == 0){
+                    throw PoiException.operateFail();
+                }
             }
         }
     }
